Test cases in wrong class. Fixing void return type
