+++ DurableFunctionsAnalyzer/DurableFunctionsAnalyzer.Test/ArgumentTypeAnalyzerTests.cs
@@ -19,215 +19,7 @@ namespace DurableFunctionsAnalyzer.Test
 
             VerifyCSharpDiagnostic(test);
         }
-        [TestMethod]
-        public void Should_not_find_any_issue_with_string_keyword()
-        {
-            var test = @"using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationTrigger] DurableOrchestrationContext context,
-            ILogger log)
-            {
-                var applications = context.GetInput<List<Application>>();
-                var approvals = await context.CallActivityAsync<string>(""ApplicationsFiltered"", new String(""));
-                log.LogInformation($""Approval received. {approvals.Count} applicants approved"");
-                return approvals.OrderByDescending(x => x.Score).First();
-            }
-
-        [FunctionName(""ApplicationsFiltered"")]
-        public static async Task<string> Run(
-            [ActivityTrigger] string userName,
-            [OrchestrationClient] DurableOrchestrationClient client)
-        {
-            await client.RaiseEventAsync(approval.InstanceId, ""ApplicationsFiltered"", approval.Applications);
-        }
-    }
-}";
-
-            VerifyCSharpDiagnostic(test);
-
-        }
-
-        [TestMethod]
-        public void Should_not_find_any_issue_with_correctly_function_parameter_but_different_long_representations()
-        {
-            var test = @"using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationTrigger] DurableOrchestrationContext context,
-            ILogger log)
-            {
-                var applications = context.GetInput<List<Application>>();
-                var approvals = await context.CallActivityAsync<List<long>>(""ApplicationsFiltered"", new String(""));
-                log.LogInformation($""Approval received. {approvals.Count} applicants approved"");
-                return approvals.OrderByDescending(x => x.Score).First();
-            }
-
-        [FunctionName(""ApplicationsFiltered"")]
-        public static async Task<List<Int64>> Run(
-            [ActivityTrigger] String userName,
-            [OrchestrationClient] DurableOrchestrationClient client)
-        {
-            await client.RaiseEventAsync(approval.InstanceId, ""ApplicationsFiltered"", approval.Applications);
-        }
-    }
-}";
-
-            VerifyCSharpDiagnostic(test);
-
-        }
-
-
-
-        [TestMethod]
-        public void Should_not_find_any_issue_with_correctly_function_parameter_but_different_int_representations()
-        {
-            var test = @"using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationTrigger] DurableOrchestrationContext context,
-            ILogger log)
-            {
-                var applications = context.GetInput<List<Application>>();
-                var approvals = await context.CallActivityAsync<List<Int>>(""ApplicationsFiltered"", new String(""));
-                log.LogInformation($""Approval received. {approvals.Count} applicants approved"");
-                return approvals.OrderByDescending(x => x.Score).First();
-            }
-
-        [FunctionName(""ApplicationsFiltered"")]
-        public static async Task<List<int>> Run(
-            [ActivityTrigger] String userName,
-            [OrchestrationClient] DurableOrchestrationClient client)
-        {
-            await client.RaiseEventAsync(approval.InstanceId, ""ApplicationsFiltered"", approval.Applications);
-        }
-    }
-}";
-
-            VerifyCSharpDiagnostic(test);
-
-        }
-
-        [TestMethod]
-        public void Should_not_find_any_issue_with_correctly_function_parameter_but_different_double_representations()
-        {
-            var test = @"using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationTrigger] DurableOrchestrationContext context,
-            ILogger log)
-            {
-                var applications = context.GetInput<List<Application>>();
-                var approvals = await context.CallActivityAsync<List<double>>(""ApplicationsFiltered"", new String(""));
-                log.LogInformation($""Approval received. {approvals.Count} applicants approved"");
-                return approvals.OrderByDescending(x => x.Score).First();
-            }
-
-        [FunctionName(""ApplicationsFiltered"")]
-        public static async Task<List<Double>> Run(
-            [ActivityTrigger] String userName,
-            [OrchestrationClient] DurableOrchestrationClient client)
-        {
-            await client.RaiseEventAsync(approval.InstanceId, ""ApplicationsFiltered"", approval.Applications);
-        }
-    }
-}";
-
-            VerifyCSharpDiagnostic(test);
-
-        }
-
-        [TestMethod]
-        public void Should_not_find_any_issue_with_correctly_function_parameter_but_different_bool_representations()
-        {
-            var test = @"using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationTrigger] DurableOrchestrationContext context,
-            ILogger log)
-            {
-                var applications = context.GetInput<List<Application>>();
-                var approvals = await context.CallActivityAsync<List<bool>>(""ApplicationsFiltered"", new String(""));
-                log.LogInformation($""Approval received. {approvals.Count} applicants approved"");
-                return approvals.OrderByDescending(x => x.Score).First();
-            }
-
-        [FunctionName(""ApplicationsFiltered"")]
-        public static async Task<List<Boolean>> Run(
-            [ActivityTrigger] String userName,
-            [OrchestrationClient] DurableOrchestrationClient client)
-        {
-            await client.RaiseEventAsync(approval.InstanceId, ""ApplicationsFiltered"", approval.Applications);
-        }
-    }
-}";
-
-            VerifyCSharpDiagnostic(test);
-
-        }
-
-
-
+   
         [TestMethod]
         public void Should_not_find_any_issue_with_correctly_function_parameter()
         {
@@ -268,6 +60,9 @@ namespace ExternalInteraction
             VerifyCSharpDiagnostic(test);
             
         }
+
+       
+
 
         [TestMethod]
         public void Should_find_issue_with_incorrect_function_parameter()
